requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
openai==1.75.0
supabase==1.2.0